src/edge_gateway/requirements.txt
paho-mqtt==2.1.0
torch==2.10.0
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.7.2
joblib==1.5.3